backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
greenlet==3.0.3
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pyotp==2.9.0
sentry-sdk==2.23.0
